fix: items not inheriting their cloned stats
