+++ Content/Items/Gear/VanillaItems/Clones/AshWoodSword.cs
@@ -1,0 +1,15 @@
+﻿using PathOfTerraria.Core;
+using Terraria.ID;
+
+namespace PathOfTerraria.Content.Items.Gear.VanillaItems.Clones;
+
+internal class AshWoodSword : VanillaClone
+{
+	protected override short VanillaItemId => ItemID.AshWoodSword;
+	
+	public override void Defaults()
+	{
+		ItemType = ItemType.Sword;
+		base.Defaults();
+	}
+}
